Move cra-ts-kitchen-sink argType stories over
