--- code/examples/cra-ts-kitchen-sink/src/stories/docgen-tests/docgen.stories.mdx
@@ -1,29 +1,0 @@
-import { Meta, ArgsTable, Description } from '@storybook/addon-docs';
-import * as DocgenJS from './DocgenJS';
-import * as DocgenTS from './DocgenTS';
-
-<Meta title="Docgen/Props" />
-
-export const DescriptionProps = ({ of }) => (
-  <>
-    <h2>{of.displayName}</h2>
-    <Description of={of} />
-    <ArgsTable of={of} />
-  </>
-);
-
-# React Docgen Test Cases
-
-<div>
-  {Object.entries(DocgenJS).map(([key, val]) => (
-    <DescriptionProps key={key} of={val} />
-  ))}
-</div>
-
-# React Typescript Docgen Test Cases
-
-<div>
-  {Object.entries(DocgenTS).map(([key, val]) => (
-    <DescriptionProps key={key} of={val} />
-  ))}
-</div>
