--- code/examples/cra-ts-kitchen-sink/src/stories/docgen-tests/types/types.stories.mdx
@@ -1,15 +1,0 @@
-import { Meta, ArgsTable } from '@storybook/addon-docs';
-import { PropTypesProps } from './prop-types';
-import { TypeScriptProps, TypeScriptHtmlComponent } from './ts-types';
-
-<Meta title="Docgen/types" />
-
-## Prop Types
-
-<ArgsTable of={PropTypesProps} />
-
-## TypeScript
-
-<ArgsTable of={TypeScriptProps} />
-
-<ArgsTable of={TypeScriptHtmlComponent} />
