+++ code/renderers/react/template/stories/__testfixtures__/js-proptypes/ext.js
