+++ code/renderers/react/template/stories/__testfixtures__/ts-html/argTypes.snapshot
@@ -1,0 +1,3 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`react component properties ts-html 1`] = `Object {}`;
