+++ code/renderers/react/template/stories/__testfixtures__/ts-html/input.tsx
@@ -1,0 +1,11 @@
+import React, { FC } from 'react';
+
+interface TypeScriptHtmlComponentProps {
+  text: string;
+}
+
+export const TypeScriptHtmlComponent: FC<
+  React.HTMLAttributes<HTMLDivElement> & TypeScriptHtmlComponentProps
+> = () => <div>My HTML component</div>;
+
+export const component = TypeScriptHtmlComponent;
