+++ code/renderers/react/template/stories/__testfixtures__/ts-html/properties.snapshot
@@ -1,0 +1,7 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`react component properties ts-html 1`] = `
+Object {
+  "rows": Array [],
+}
+`;
