+++ code/renderers/react/template/stories/__testfixtures__/ts-jsdoc/argTypes.snapshot
@@ -1,0 +1,3 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`react component properties ts-jsdoc 1`] = `Object {}`;
