+++ code/renderers/react/template/stories/__testfixtures__/ts-jsdoc/docgen.snapshot
@@ -1,0 +1,23 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`react component properties ts-jsdoc 1`] = `
+"import React from 'react';
+
+/**
+ * Button functional component (React.FC)
+ */
+const TypeScriptProps = ({
+  case1,
+  case2
+}) => /*#__PURE__*/React.createElement(\\"button\\", {
+  type: \\"button\\",
+  onClick: case1
+}, \\"JSDoc with TypeScript! \\", case2);
+
+export const component = TypeScriptProps;
+TypeScriptProps.__docgenInfo = {
+  \\"description\\": \\"Button functional component (React.FC)\\",
+  \\"methods\\": [],
+  \\"displayName\\": \\"TypeScriptProps\\"
+};"
+`;
