+++ code/renderers/react/template/stories/__testfixtures__/ts-types/input.tsx
@@ -126,10 +126,4 @@ TypeScriptProps.defaultProps = {
   inlinedNumericLiteralUnion: 1,
 };
 
-interface TypeScriptHtmlComponentProps {
-  text: string;
-}
-
-export const TypeScriptHtmlComponent: FC<
-  React.HTMLAttributes<HTMLDivElement> & TypeScriptHtmlComponentProps
-> = () => <div>My HTML component</div>;
+export const component = TypeScriptProps;
